+++ data/normalized/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Latest Microsoft Security News Changes
 
-Generated: 2026-04-02T07:08:36.794Z
+Generated: 2026-04-03T07:04:37.440Z
 
 - 2026-07-01 • Microsoft Sentinel • Action required • Call to action: update queries and automation by July 1, 2026 - standardized account entity naming in incidents and alerts
 - 2026-07-01 • Microsoft Sentinel • GA • , Microsoft Sentinel will be supported in the Defender portal only, and any remaining customers using the Azure portal will be automatically redirected.
@@ -9,6 +9,8 @@ Generated: 2026-04-02T07:08:36.794Z
 - 2026-06-01 • Microsoft Entra • Action required • General Availability - Sign-in with username/alias
 - 2026-06-01 • Microsoft Sentinel • Action required • Required action
 - 2026-04-01 • Microsoft Defender for Cloud • GA • General availability of container security capabilities in Azure Government cloud
+- 2026-04-01 • Microsoft Defender for Endpoint • Preview • Release - macOS
+- 2026-04-01 • Microsoft Defender for Endpoint • Ikke oppgitt • Release version 20.126022.18.0 released: See
 - 2026-04-01 • Microsoft Purview • GA • General availability (GA)
 - 2026-03-31 • Microsoft Defender for Cloud • Ikke oppgitt • Malware automated remediation in Defender for Storage
 - 2026-03-29 • Microsoft Defender for Cloud • Preview • AI model security for Azure Machine Learning (Preview)
@@ -20,5 +22,3 @@ Generated: 2026-04-02T07:08:36.794Z
 - 2026-03-04 • Microsoft Defender for Cloud • Preview • On-demand malware scanning of Azure Files in Microsoft Defender for Storage
 - 2026-03-01 • Microsoft Defender for Endpoint • GA • Library management for live response is now generally available. This feature provides a centralized view for managing files and scripts used during live response sessions.
 - 2026-03-01 • Microsoft Defender for Endpoint • Preview • New Microsoft Secure Score recommendations
-- 2026-03-01 • Microsoft Defender for Endpoint • Preview • Microsoft Secure Score now includes new recommendations to help organizations proactively prevent common endpoint attack techniques
-- 2026-03-01 • Microsoft Defender for Endpoint • Ikke oppgitt • Block outbound network connections from Microsoft HTML Application Host (mshta.exe)